Moving Line into a folder and ErrorBar function into one separate file
